+++ packages/backend/src/core/sse/sse.service.ts
@@ -1,10 +1,10 @@
 import { DockerService } from '@/modules/docker/docker.service';
 import { colorizeLogs } from '@/modules/docker/helpers/colorize-logs';
 import { Injectable, type MessageEvent } from '@nestjs/common';
+import type { SSE, Topic } from '@runtipi/common/schemas';
 import type { AppUrn } from '@runtipi/common/types';
 import { Observable, Subject, interval } from 'rxjs';
 import { LoggerService } from '../logger/logger.service';
-import type { SSE, Topic } from './dto/sse.dto';
 
 @Injectable()
 export class SSEService {
