--- packages/backend/src/exports/index.ts
@@ -1,3 +1,0 @@
-import { type SSE, type Topic, sseSchema } from '../core/sse/dto/sse.dto';
-
-export { sseSchema, type Topic, type SSE };
