+++ packages/common/src/schemas/index.ts
@@ -4,6 +4,8 @@ import type { DependsOn, DynamicCompose, Service, ServiceInput } from './dynamic
 import { APP_CATEGORIES, ARCHITECTURES, FIELD_TYPES, RANDOM_ENCODINGS, appInfoSchema, formFieldSchema } from './app-info.js';
 import type { AppCategory, AppInfo, AppInfoInput, FieldType, FormField, RandomEncoding } from './app-info.js';
 
+import { type SSE, type Topic, sseSchema } from './sse.js';
+
 import { toJsonSchema } from './utils/to-json-schema.js';
 
 export {
@@ -16,6 +18,7 @@ export {
   FIELD_TYPES,
   ARCHITECTURES,
   appInfoSchema,
+  sseSchema,
   type ServiceInput,
   type DependsOn,
   type Service,
@@ -26,4 +29,6 @@ export {
   type FieldType,
   type RandomEncoding,
   type AppCategory,
+  type SSE,
+  type Topic,
 };
