+++ packages/frontend/package.json
@@ -30,7 +30,6 @@
     "@tanstack/react-query": "^5.75.7",
     "@tanstack/react-query-devtools": "^5.75.7",
     "@uidotdev/usehooks": "^2.4.1",
-    "backend": "workspace:*",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "dompurify": "^3.2.5",
@@ -77,6 +76,7 @@
     "globals": "^16.1.0",
     "typescript": "^5.8.3",
     "vite": "^6.3.5",
+    "vite-tsconfig-paths": "5.1.4",
     "vitest": "^3.1.3"
   }
 }
